Refactor TypefaceHelper.java; se crea CalligraphyTypefaceSetter.java y se mueven test especificos a otra clase. Se agrega metodo TypefaceHelper.getFontTypeface para obtener una typeface sin enviar un callback
