+++ ui/src/main/java/com/mercadolibre/android/ui/font/CalligraphyTypefaceSetter.java
@@ -1,0 +1,38 @@
+package com.mercadolibre.android.ui.font;
+
+import android.content.Context;
+import android.graphics.Paint;
+import android.graphics.Typeface;
+import android.support.annotation.NonNull;
+import android.widget.Switch;
+import android.widget.TextView;
+
+import uk.co.chrisjenx.calligraphy.TypefaceUtils;
+
+public class CalligraphyTypefaceSetter implements TypefaceHelper.TypefaceSetter {
+    @Override
+    public <T extends TextView> void setTypeface(@NonNull final T view, @NonNull final Font font) {
+        final Typeface typeface = createTypeface(view.getContext(), font);
+        view.setTypeface(typeface);
+
+        if (view instanceof Switch) {
+            ((Switch) view).setSwitchTypeface(typeface);
+        }
+    }
+
+    @Override
+    public void setTypeface(@NonNull final Context context, @NonNull final Paint paint, @NonNull final Font font) {
+        final Typeface typeface = createTypeface(context, font);
+        paint.setTypeface(typeface);
+    }
+
+    @Override
+    public Typeface getTypeface(@NonNull Context context, @NonNull Font font) {
+        return createTypeface(context, font);
+    }
+
+    private Typeface createTypeface(@NonNull Context context, @NonNull Font font) {
+        return TypefaceUtils.load(context.getAssets(), font.getFontPath());
+    }
+
+}
