+++ ui/src/main/java/com/mercadolibre/android/ui/font/TypefaceHelper.java
@@ -4,13 +4,8 @@ import android.content.Context;
 import android.graphics.Paint;
 import android.graphics.Typeface;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.support.v4.content.res.ResourcesCompat;
-import android.support.v4.provider.FontsContractCompat;
-import android.widget.Switch;
 import android.widget.TextView;
-
-import uk.co.chrisjenx.calligraphy.TypefaceUtils;
 
 /**
  * This class is used as a wrapper for our custom font.
@@ -22,43 +17,7 @@ public final class TypefaceHelper {
     private static TypefaceSetter typefaceSetter;
 
     static {
-        typefaceSetter = new TypefaceSetter() {
-            @Override
-            public <T extends TextView> void setTypeface(@NonNull final T view, @NonNull final Font font) {
-                final Typeface typeface = createTypeface(view.getContext(), font);
-                view.setTypeface(typeface);
-
-                if (view instanceof Switch) {
-                    ((Switch) view).setSwitchTypeface(typeface);
-                }
-            }
-
-            @Override
-            public void setTypeface(@NonNull final Context context, @NonNull final Paint paint, @NonNull final Font font) {
-                final Typeface typeface = createTypeface(context, font);
-                paint.setTypeface(typeface);
-            }
-
-            @Override
-            @SuppressWarnings("PMD.LinguisticNaming")
-            public void getTypeface(@NonNull final Context context, @NonNull final Font font, @NonNull final ResourcesCompat.FontCallback fontCallback) {
-                final Typeface typeface = createTypeface(context, font);
-
-                if (typeface == null) {
-                    fontCallback.onFontRetrievalFailed(FontsContractCompat.FontRequestCallback.FAIL_REASON_FONT_NOT_FOUND);
-                } else {
-                    fontCallback.onFontRetrieved(typeface);
-                }
-            }
-
-            @Nullable
-            private Typeface createTypeface(@NonNull Context context, @Nullable Font font) {
-                if (font == null) {
-                    return null;
-                }
-                return TypefaceUtils.load(context.getAssets(), font.getFontPath());
-            }
-        };
+        typefaceSetter = new CalligraphyTypefaceSetter();
     }
 
     private TypefaceHelper() {
@@ -101,10 +60,25 @@ public final class TypefaceHelper {
      * @param context to use
      * @param font to retrieve its typeface
      * @param fontCallback to call when the typeface is retrieved
+     *
+     * @deprecated use geyFontTypeface instead
      */
     @SuppressWarnings("PMD.LinguisticNaming")
+    @Deprecated
     public static void getTypeface(@NonNull final Context context, @NonNull final Font font, @NonNull final ResourcesCompat.FontCallback fontCallback) {
-        typefaceSetter.getTypeface(context, font, fontCallback);
+        Typeface typeface = typefaceSetter.getTypeface(context, font);
+        fontCallback.onFontRetrieved(typeface);
+    }
+
+    /**
+     * Get a tyoeface associated to the font passed.
+
+     * @param context to use
+     * @param font to use
+     * @return associated typeface
+     */
+    public static Typeface geyFontTypeface(@NonNull final Context context, @NonNull Font font) {
+        return typefaceSetter.getTypeface(context, font);
     }
 
     /**
@@ -129,16 +103,11 @@ public final class TypefaceHelper {
         void setTypeface(@NonNull final Context context, @NonNull final Paint paint, @NonNull final Font font);
 
         /**
-         * Get a typeface associated to the font passed. The typeface will be sent through the
-         * font callback passed as param
-         *
+         * Return the typeface associated with the font
          * @param context to use
-         * @param font to retrieve its typeface
-         * @param fontCallback to call when the typeface is retrieved
+         * @param font to find the typeface
+         * @return typeface associated
          */
-        @SuppressWarnings("PMD.LinguisticNaming")
-        void getTypeface(@NonNull final Context context, @NonNull final Font font, @NonNull final ResourcesCompat.FontCallback fontCallback);
-
+        Typeface getTypeface(@NonNull Context context, @NonNull final Font font);
     }
-
 }
